Move core parsing logic into ear-core
